refactor(verify): extract cwd resolution
